requirements.txt
slack-bolt
langchain
langchain-google-community
openai
python-dotenv
chromadb

